05_parallel_for example changed to Fortran main (C++ main seems hard to compile with fortran kernels for some reason ---to be investigated).
