+++ core/fortran-example/tutorial/05_parallel_for/CMakeLists.txt
@@ -1,3 +1,7 @@
-  add_executable(ForParallelFor parallel_for_example.F90 main.cpp)
+  add_executable(ForParallelFor Fortran_main.f90 Fortran_kernels.F90 parallel_for_example.cpp)
+  set(CMAKE_Fortran_FLAGS "-ta=tesla:cuda9.0 -Minfo=accel -Mcuda=cuda9.0")
+  set(CMAKE_CXX_FLAGS "--relocatable-device-code=true")
+  set(CMAKE_EXE_LINKER_FLAGS "-pgc++libs")
   target_link_libraries(ForParallelFor cabanacore)
-  add_test(NAME Fortran_tutorial_11 COMMAND ForParallelFor)
+  add_test(NAME Fortran_tutorial_06 COMMAND ForParallelFor)
+  set_property(TARGET ForParallelFor PROPERTY LINKER_LANGUAGE Fortran)
